+++ es6/doc-utils.js
@@ -382,6 +382,7 @@ const corruptCharacters = /[\x00-\x08\x0B\x0C\x0E-\x1F]/g;
  * 1F    US  (unit separator)
  */
 function hasCorruptCharacters(string) {
+	corruptCharacters.lastIndex = 0;
 	return corruptCharacters.test(string);
 }